minor improvement for priv level handling
