+++ scrapli/driver/core/cisco_iosxr/driver.py
@@ -22,6 +22,7 @@ def iosxr_on_open(conn: NetworkDriver) -> None:
     # sleep for session to establish; without this we never find base prompt for some reason?
     # maybe this is an artifact from previous iterations/tests and can be done away with...
     time.sleep(1)
+    conn.acquire_priv(conn.default_desired_priv)
     conn.channel.send_inputs("terminal length 0")
